feat: add datatypes
