+++ pkg/report/output/dataflow/dataflow_test.go
@@ -1,32 +1,32 @@
-package output_test
+package dataflow_test
 
 import (
 	"os"
 	"testing"
 
-	"github.com/bearer/curio/pkg/report/output"
+	"github.com/bearer/curio/pkg/report/output/dataflow"
 	"github.com/bearer/curio/pkg/types"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestDataflow(t *testing.T) {
+func TestDataType(t *testing.T) {
 	testCases := []struct {
 		Name        string
 		FileContent string
-		Want        *output.DataFlow
+		Want        *dataflow.DataFlow
 	}{
 		{
 			Name:        "single detection",
 			FileContent: `{"type": "schema", "detector_type":"ruby", "source": {"filename": "./users.rb", "line_number": 25}, "value": {"field_name": "User"}}`,
-			Want: &output.DataFlow{
-				Datatypes: []output.Datatype{
+			Want: &dataflow.DataFlow{
+				Datatypes: []dataflow.Datatype{
 					{
 						Name: "User",
-						Detectors: []output.Detector{
+						Detectors: []dataflow.Detector{
 							{
 								Name:   "ruby",
 								Stored: true,
-								Locations: []output.Location{
+								Locations: []dataflow.Location{
 									{Filename: "./users.rb", LineNumber: 25},
 								},
 							},
@@ -39,15 +39,15 @@ func TestDataflow(t *testing.T) {
 			Name: "single detection - duplicates",
 			FileContent: `{"type": "schema", "detector_type":"ruby", "source": {"filename": "./users.rb", "line_number": 25}, "value": {"field_name": "User"}}
 {"type": "schema", "detector_type":"ruby", "source": {"filename": "./users.rb", "line_number": 25}, "value": {"field_name": "User"}}`,
-			Want: &output.DataFlow{
-				Datatypes: []output.Datatype{
+			Want: &dataflow.DataFlow{
+				Datatypes: []dataflow.Datatype{
 					{
 						Name: "User",
-						Detectors: []output.Detector{
+						Detectors: []dataflow.Detector{
 							{
 								Name:   "ruby",
 								Stored: true,
-								Locations: []output.Location{
+								Locations: []dataflow.Location{
 									{Filename: "./users.rb", LineNumber: 25},
 								},
 							},
@@ -60,15 +60,15 @@ func TestDataflow(t *testing.T) {
 			Name: "single detection - with wierd data in report",
 			FileContent: `{"type": "schema", "detector_type":"ruby", "source": {"filename": "./users.rb", "line_number": 25}, "value": {"field_name": "User"}}
 {"user": true }`,
-			Want: &output.DataFlow{
-				Datatypes: []output.Datatype{
+			Want: &dataflow.DataFlow{
+				Datatypes: []dataflow.Datatype{
 					{
 						Name: "User",
-						Detectors: []output.Detector{
+						Detectors: []dataflow.Detector{
 							{
 								Name:   "ruby",
 								Stored: true,
-								Locations: []output.Location{
+								Locations: []dataflow.Location{
 									{Filename: "./users.rb", LineNumber: 25},
 								},
 							},
@@ -81,22 +81,22 @@ func TestDataflow(t *testing.T) {
 			Name: "multiple detections - with same object name - deterministic output",
 			FileContent: `{"type": "schema", "detector_type":"ruby", "source": {"filename": "./users.rb", "line_number": 25}, "value": {"field_name": "User"}}
 			{"type": "schema", "detector_type":"csharp", "source": {"filename": "./users.cs", "line_number": 12}, "value": {"field_name": "User"}}`,
-			Want: &output.DataFlow{
-				Datatypes: []output.Datatype{
+			Want: &dataflow.DataFlow{
+				Datatypes: []dataflow.Datatype{
 					{
 						Name: "User",
-						Detectors: []output.Detector{
+						Detectors: []dataflow.Detector{
 							{
 								Name:   "csharp",
 								Stored: true,
-								Locations: []output.Location{
+								Locations: []dataflow.Location{
 									{Filename: "./users.cs", LineNumber: 12},
 								},
 							},
 							{
 								Name:   "ruby",
 								Stored: true,
-								Locations: []output.Location{
+								Locations: []dataflow.Location{
 									{Filename: "./users.rb", LineNumber: 25},
 								},
 							},
@@ -109,15 +109,15 @@ func TestDataflow(t *testing.T) {
 			Name: "multiple detections - with different names",
 			FileContent: `{"type": "schema", "detector_type":"ruby", "source": {"filename": "./users.rb", "line_number": 25}, "value": {"field_name": "User"}}
 			{"type": "schema", "detector_type":"csharp", "source": {"filename": "./users.cs", "line_number": 12}, "value": {"field_name": "Address"}}`,
-			Want: &output.DataFlow{
-				Datatypes: []output.Datatype{
+			Want: &dataflow.DataFlow{
+				Datatypes: []dataflow.Datatype{
 					{
 						Name: "Address",
-						Detectors: []output.Detector{
+						Detectors: []dataflow.Detector{
 							{
 								Name:   "csharp",
 								Stored: true,
-								Locations: []output.Location{
+								Locations: []dataflow.Location{
 									{Filename: "./users.cs", LineNumber: 12},
 								},
 							},
@@ -125,11 +125,11 @@ func TestDataflow(t *testing.T) {
 					},
 					{
 						Name: "User",
-						Detectors: []output.Detector{
+						Detectors: []dataflow.Detector{
 							{
 								Name:   "ruby",
 								Stored: true,
-								Locations: []output.Location{
+								Locations: []dataflow.Location{
 									{Filename: "./users.rb", LineNumber: 25},
 								},
 							},
@@ -155,7 +155,7 @@ func TestDataflow(t *testing.T) {
 			}
 			file.Close()
 
-			dataflow, err := output.GetDataFlowOutput(types.Report{
+			dataflow, err := dataflow.GetDataFlowOutput(types.Report{
 				Path: file.Name(),
 			})
 			if err != nil {
