+++ pkg/report/output/dataflow/datatypes/datatypes.go
@@ -1,4 +1,4 @@
-package output
+package datatypes
 
 import (
 	"bytes"
@@ -7,86 +7,8 @@ import (
 
 	"github.com/bearer/curio/pkg/report/detections"
 	"github.com/bearer/curio/pkg/report/schema"
-	"github.com/bearer/curio/pkg/types"
 	"github.com/bearer/curio/pkg/util/maputil"
-	"github.com/rs/zerolog/log"
 )
-
-type DataFlow struct {
-	Datatypes []Datatype `json:"data_types,omitempty"`
-	Risks     []Datatype `json:"risks,omitempty"`
-}
-
-type Datatype struct {
-	Name      string     `json:"name"`
-	Detectors []Detector `json:"detectors"`
-}
-
-type Detector struct {
-	Name      string     `json:"name"`
-	Stored    bool       `json:"stored"`
-	Locations []Location `json:"locations"`
-}
-
-type Location struct {
-	Filename   string `json:"filename"`
-	LineNumber int    `json:"line_number"`
-}
-
-var dataflowDetections []detections.DetectionType = []detections.DetectionType{detections.TypeSchema, detections.TypeCustom, detections.TypeSchemaClassified}
-
-func GetDataFlowOutput(report types.Report) (*DataFlow, error) {
-	holder := dataFlowHolder{
-		datatypes: make(map[string]*datatypeHolder),
-	}
-
-	reportedDetections, err := GetDetectorsOutput(report)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, detection := range reportedDetections {
-		detection, ok := detection.(map[string]interface{})
-		if !ok {
-			return nil, fmt.Errorf("found detection in report which is not object")
-		}
-
-		detectionType, ok := detection["type"].(string)
-
-		isDataflow := false
-		for _, allowedDetection := range dataflowDetections {
-			if detections.DetectionType(detectionType) == allowedDetection {
-				isDataflow = true
-			}
-		}
-
-		if !ok || !isDataflow {
-			continue
-		}
-
-		var castedDetection detections.Detection
-		buf := bytes.NewBuffer(nil)
-		err := json.NewEncoder(buf).Encode(detection)
-		if err != nil {
-			return nil, err
-		}
-		err = json.NewDecoder(buf).Decode(&castedDetection)
-		if err != nil {
-			return nil, err
-		}
-
-		err = holder.addSchema(castedDetection)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return holder.toDataFlow(), nil
-}
-
-type dataFlowHolder struct {
-	datatypes map[string]*datatypeHolder // group datatypeHolders by name
-}
 
 type datatypeHolder struct {
 	name      string
@@ -104,7 +26,6 @@ type fileHolder struct {
 func (holder *dataFlowHolder) toDataFlow() *DataFlow {
 	data := &DataFlow{}
 
-	log.Debug().Msgf("datatypes are %#v", holder.datatypes)
 	datatypes := maputil.ToSortedSlice(holder.datatypes)
 
 	for _, datatype := range datatypes {
